Crate simple_redis

simple_redis

Simple and resilient redis client based on redis-rs with internal connection and subscription handling.

This library provides a very basic, simple API for the most common redis operations.
While not as comprehensive or flexiable as redis-rs, it does provide a simpler api for most common use cases and operations as well as automatic and resilient internal connection and subscription (pubsub) handling.
In addition, the entire API is accessible via redis client and there is no need to manage connection or pubsub instances in parallel.

Connection Resiliency

Connection resiliency is managed by verifying the internally managed connection before every operation against the redis server.
In case of any connection issue, a new connection will be allocated to ensure the operation is invoked on a valid connection only.
However, this comes at a small performance cost of PING operation to the redis server.

Subscription Resiliency

Subscription resiliency is ensured by recreating the internal pubsub and issuing new subscription requests automatically in case of any error while fetching a message from the subscribed channels.

Examples

Initialization and Simple Operations

extern crate simple_redis;

fn main() {
    match simple_redis::create("redis://127.0.0.1:6379/") {
        Ok(mut client) =>  {
            println!("Created Redis Client");

            match client.set("my_key", "my_value") {
                Err(error) => println!("Unable to set value in Redis: {}", error),
                _ => println!("Value set in Redis")
            };

            match client.get_string("my_key") {
                Ok(value) => println!("Read value from Redis: {}", value),
                Err(error) => println!("Unable to get value from Redis: {}", error)
            };

            match client.set("my_numeric_key", 255.5) {
                Err(error) => println!("Unable to set value in Redis: {}", error),
                _ => println!("Value set in Redis")
            };

            match client.get::<f32>("my_numeric_key") {
                Ok(value) => println!("Read value from Redis: {}", value),
                Err(error) => println!("Unable to get value from Redis: {}", error)
            };

            match client.hgetall("my_map") {
                Ok(map) => {
                    match map.get("my_field") {
                        Some(value) => println!("Got field value from map: {}", value),
                        None => println!("Map field is emtpy"),
                    }
                },
                Err(error) => println!("Unable to read map from Redis: {}", error),
            };

            /// run some command that is not built in the library
            match client.run_command::<String>("ECHO", vec!["testing"]) {
                Ok(value) => assert_eq!(value, "testing"),
                _ => panic!("test error"),
            };

            /// publish messages
            let result = client.publish("news_channel", "test message");
            assert!(result.is_ok());
        },
        Err(error) => println!("Unable to create Redis client: {}", error)
    }
}

Subscription Flow

extern crate simple_redis;

fn main() {
    match simple_redis::create("redis://127.0.0.1:6379/") {
        Ok(mut client) =>  {
            println!("Created Redis Client");

            let mut result = client.subscribe("important_notifications");
            assert!(result.is_ok());
            result = client.psubscribe("*_notifications");
            assert!(result.is_ok());

            loop {
                // fetch next message (wait up to 5 seconds, 0 for no timeout)
                match client.get_message(5000) {
                    Ok(message) => {
                        let payload: String = message.get_payload().unwrap();
                        assert_eq!(payload, "my important message")
                    },
                    Err(error) => println!("Error while fetching message, should retry again, info: {}", error),
                }
            }
        },
        Err(error) => println!("Unable to create Redis client: {}", error)
    }
}

Installation

In order to use this library, just add it as a dependency:

[dependencies]
simple_redis = "*"

Contributing

See contributing guide

License

Developed by Sagie Gur-Ari and licensed under the Apache 2 open source license.

Modules

client

types

Functions

create
Constructs a new redis client.
The redis connection string must be in the following format: redis://[:<passwd>@]<hostname>[:port][/<db>]

Type Definitions

ErrorInfo
Error Info

Message
PubSub message

RedisError
Error Type

RedisResult
Redis result which either holds a value or a Redis error