Module simple_redis::types

types

Defines the various types and aliases used or exposed by the simple_redis library.

Structs

RedisError
Redis Error struct

Enums

ErrorInfo
Holds the error information

Traits

RedisArg
Defines a redis command argument

Type Definitions

Message
PubSub message

RedisBoolResult
Holds bool result or error

RedisEmptyResult
Holds empty result or error

RedisMessageResult
Holds pubsub message result or error

RedisResult
Redis result which either holds a value or a Redis error

RedisStringResult
Holds string result or error